+++ Model/Content/Converter/CmsEntityConverterManager.php
@@ -6,7 +6,6 @@ namespace Overdose\CMSContent\Model\Content\Converter;
 
 use Magento\Framework\Exception\LocalizedException;
 use Overdose\CMSContent\Api\CmsEntityConverterManagerInterface;
-use function Overdose\CMSContent\Model\Converter\__;
 
 class CmsEntityConverterManager implements CmsEntityConverterManagerInterface
 {