+++ Model/Content/Generator/CmsEntityGeneratorManager.php
@@ -6,7 +6,6 @@ namespace Overdose\CMSContent\Model\Content\Generator;
 
 use Magento\Framework\Exception\LocalizedException;
 use Overdose\CMSContent\Api\CmsEntityGeneratorManagerInterface;
-use function Overdose\CMSContent\Model\Generator\__;
 
 class CmsEntityGeneratorManager implements CmsEntityGeneratorManagerInterface
 {