+++ Model/Content/Import.php
@@ -1,15 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Overdose\CMSContent\Model\Content;
 
-use Magento\Cms\Api\BlockRepositoryInterface;
+use Exception;
 use Magento\Cms\Api\Data\BlockInterface as CmsBlockInterface;
 use Magento\Cms\Api\Data\PageInterface as CmsPageInterface;
-use Magento\Cms\Api\PageRepositoryInterface;
-use Magento\Cms\Model\BlockFactory as CmsBlockFactory;
-use Magento\Cms\Model\PageFactory as CmsPageFactory;
-use Magento\Cms\Model\ResourceModel\Block\CollectionFactory as CmsBlockCollectionFactory;
-use Magento\Cms\Model\ResourceModel\Page\CollectionFactory as CmsPageCollectionFactory;
+use Magento\Framework\Api\SearchCriteriaBuilder;
+use Magento\Framework\Exception\FileSystemException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Filesystem\Io\File;
 use Magento\Framework\Serialize\SerializerInterface;
@@ -18,6 +17,7 @@ use Overdose\CMSContent\Api\ContentVersionManagementInterface;
 use Overdose\CMSContent\Api\StoreManagementInterface;
 use Overdose\CMSContent\File\FileManagerInterface;
 use Overdose\CMSContent\Model\Config;
+use Overdose\CMSContent\Model\EntityManagement;
 
 class Import implements ContentImportInterface
 {
@@ -47,31 +47,6 @@ class Import implements ContentImportInterface
     private $mediaMode = ContentImportInterface::OD_MEDIA_MODE_UPDATE;
 
     /**
-     * @var CmsPageCollectionFactory
-     */
-    private $pageCollectionFactory;
-
-    /**
-     * @var CmsBlockCollectionFactory
-     */
-    private $blockCollectionFactory;
-
-    /**
-     * @var CmsPageFactory
-     */
-    private $pageFactory;
-
-    /**
-     * @var CmsBlockFactory
-     */
-    private $blockFactory;
-
-    /**
-     * @var BlockRepositoryInterface
-     */
-    private $blockRepository;
-
-    /**
      * @var StoreManagementInterface
      */
     private $storeManagement;
@@ -87,81 +62,72 @@ class Import implements ContentImportInterface
     private $contentVersionManagement;
 
     /**
-     * @var PageRepositoryInterface
+     * @var EntityManagement
      */
-    private $pageRepository;
+    private $entityManagement;
+
+    /**
+     * @var SearchCriteriaBuilder
+     */
+    private $searchCriteriaBuilder;
 
     /**
      * @param FileManagerInterface $fileManager
      * @param Config $config
      * @param File $file
-     * @param CmsPageFactory $pageFactory
-     * @param CmsBlockFactory $blockFactory
-     * @param CmsBlockCollectionFactory $blockCollectionFactory
-     * @param CmsPageCollectionFactory $pageCollectionFactory
-     * @param BlockRepositoryInterface $blockRepository
-     * @param PageRepositoryInterface $pageRepository
      * @param StoreManagementInterface $storeManagement
      * @param ContentVersionManagementInterface $contentVersionManagement
      * @param SerializerInterface $serializerInterface
+     * @param EntityManagement $entityManagement
+     * @param SearchCriteriaBuilder $searchCriteriaBuilder
      */
     public function __construct(
-        FileManagerInterface              $fileManager,
-        Config                            $config,
-        File                              $file,
-        CmsPageFactory                    $pageFactory,
-        CmsBlockFactory                   $blockFactory,
-        CmsBlockCollectionFactory         $blockCollectionFactory,
-        CmsPageCollectionFactory          $pageCollectionFactory,
-        BlockRepositoryInterface          $blockRepository,
-        PageRepositoryInterface          $pageRepository,
-        StoreManagementInterface          $storeManagement,
+        FileManagerInterface $fileManager,
+        Config $config,
+        File $file,
+        StoreManagementInterface $storeManagement,
         ContentVersionManagementInterface $contentVersionManagement,
-        SerializerInterface               $serializerInterface
+        SerializerInterface $serializerInterface,
+        EntityManagement $entityManagement,
+        SearchCriteriaBuilder $searchCriteriaBuilder
     ) {
         $this->fileManager = $fileManager;
         $this->config = $config;
         $this->file = $file;
-        $this->pageFactory = $pageFactory;
-        $this->blockFactory = $blockFactory;
-        $this->blockCollectionFactory = $blockCollectionFactory;
-        $this->pageCollectionFactory = $pageCollectionFactory;
-        $this->blockRepository = $blockRepository;
-        $this->pageRepository = $pageRepository;
         $this->storeManagement = $storeManagement;
         $this->serializerInterface = $serializerInterface;
         $this->contentVersionManagement = $contentVersionManagement;
+        $this->entityManagement = $entityManagement;
+        $this->searchCriteriaBuilder = $searchCriteriaBuilder;
     }
 
     /**
-     * Import contents from zip archive and return number of imported records (-1 on error)
-     * @param string $fileName
-     * @param bool $rm
-     * @return int
-     * @throws \Exception
+     * @inheritdoc
      */
     public function importContentFromZipFile(string $fileName, bool $rm): int
     {
         $zipArchive = new \ZipArchive();
         $res = $zipArchive->open($fileName);
         if ($res !== true) {
-            throw new \Exception('Cannot open ZIP archive');
+            throw new Exception('Cannot open ZIP archive');
         }
 
         $subPath = md5(date(DATE_RFC2822));
-        $extractPath = $this->fileManager->getFolder($this->config->getExtractPath() . DIRECTORY_SEPARATOR . $subPath);
+        $extractPath = $this->fileManager->getFolder(
+            $this->config->getExtractPath() . DIRECTORY_SEPARATOR . $subPath
+        );
 
         $zipArchive->extractTo($extractPath);
         $zipArchive->close();
 
         $count = 0;
-        foreach (scandir($extractPath. '/') as $path){
+        foreach (scandir($extractPath. '/') as $path) {
             $absolutePath = $extractPath. '/' . $path;
-            if(in_array($path, ['.', '..']) || is_dir($path)) {
+            if (in_array($path, ['.', '..']) || is_dir($path)) {
                 continue;
             }
             if (!$this->file->fileExists($absolutePath, true)) {
-                throw new \Exception($path . ' is missing');
+                throw new Exception($path . ' is missing');
             }
 
             $pathInfo = $this->file->getPathInfo($absolutePath);
@@ -188,18 +154,40 @@ class Import implements ContentImportInterface
             }
         }
 
-        // Clear archive
+        // Clear extracted folder
         $this->file->rmdir($extractPath, true);
 
         return $count;
     }
 
     /**
+     * @inheritdoc
+     */
+    public function setCmsModeOption($mode): ContentImportInterface
+    {
+        $this->cmsMode = $mode;
+
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setMediaModeOption($mode): ContentImportInterface
+    {
+        $this->mediaMode = $mode;
+
+        return $this;
+    }
+
+    /**
      * Import contents from array and return number of imported records (-1 on error)
+     *
      * @param array $payload
      * @param string|null $archivePath = null
+     *
      * @return int
-     * @throws \Exception
+     * @throws Exception
      */
     public function importContentFromArray(array $payload, string $archivePath = null): int
     {
@@ -207,13 +195,13 @@ class Import implements ContentImportInterface
             $payload = $payload['config'];
         }
         if (!isset($payload['pages']) && !isset($payload['blocks'])) {
-            throw new \Exception('Invalid import archive');
+            throw new Exception('Invalid import archive');
         }
 
         $count = 0;
 
         // Import pages
-        if(isset($payload['pages'])){
+        if (isset($payload['pages'])) {
             foreach ($payload['pages'] as $key => $pageData) {
                 if ($this->importPageContentFromArray($pageData)) {
                     $count++;
@@ -221,7 +209,7 @@ class Import implements ContentImportInterface
             }
         }
 
-        if(isset($payload['blocks'])){
+        if (isset($payload['blocks'])) {
             foreach ($payload['blocks'] as $key => $blockData) {
                 if ($this->importBlockContentFromArray($blockData)) {
                     $count++;
@@ -229,55 +217,29 @@ class Import implements ContentImportInterface
             }
         }
 
-        // Import media
-        if ($archivePath && ($count > 0) && ($this->mediaMode != ContentImportInterface::OD_MEDIA_MODE_NONE)) {
-            if (isset($payload['media'])) {
-                foreach ($payload['media'] as $mediaFile) {
-                    $sourceFile = $archivePath . '/' . self::MEDIA_ARCHIVE_PATH . '/' . $mediaFile;
-                    $destFile = $this->fileManager->getMediaPath($mediaFile);
-
-                    if ($this->file->fileExists($sourceFile, true)) {
-                        if ($this->file->fileExists($destFile, true) &&
-                            ($this->mediaMode == ContentImportInterface::OD_MEDIA_MODE_SKIP)
-                        ) {
-                            continue;
-                        }
-
-                        if (!$this->file->fileExists(dirname($destFile), false)) {
-                            if (!$this->file->mkdir(dirname($destFile))) {
-                                throw new \Exception('Unable to create folder: ' . dirname($destFile));
-                            }
-                        }
-                        if (!$this->file->cp($sourceFile, $destFile)) {
-                            throw new \Exception('Unable to save image: ' . $mediaFile);
-                        }
-                        $count++;
-                    }
-                }
-            }
-        }
-
-        return $count;
+        return $this->importMedia($payload, $archivePath, $count);
     }
 
     /**
      * Import a single page from an array and return false on error and true on success
      * @param array $pageData
      * @return bool
-     * @throws \Exception
+     * @throws Exception
      */
     public function importPageContentFromArray(array $pageData): bool
     {
-        // Process block identifiers
         $pageData = $this->loadBlocksByIdent($pageData, CmsPageInterface::IDENTIFIER);
-
-        // Will not use repositories to save pages because it does not allow stores selection
 
         $storeIds = $this->storeManagement->getStoreIdsByCodes($pageData['stores']);
 
-        $collection = $this->pageCollectionFactory->create();
-        $collection
-            ->addFieldToFilter(CmsPageInterface::IDENTIFIER, $pageData['cms'][CmsPageInterface::IDENTIFIER]);
+        $collection = $this->entityManagement
+            ->getRepository(EntityManagement::TYPE_PAGE)
+            ->getList(
+                $this->searchCriteriaBuilder->addFilter(
+                    CmsBlockInterface::IDENTIFIER,
+                    $pageData['cms'][CmsBlockInterface::IDENTIFIER]
+                )->create()
+            )->getItems();
 
         $pageId = 0;
         foreach ($collection as $item) {
@@ -291,7 +253,7 @@ class Import implements ContentImportInterface
             }
         }
 
-        $page = $this->pageFactory->create();
+        $page = $this->entityManagement->getFactory(EntityManagement::TYPE_PAGE);
         if ($pageId) {
             $page->load($pageId);
 
@@ -320,7 +282,7 @@ class Import implements ContentImportInterface
             ->setIsActive($cms[CmsPageInterface::IS_ACTIVE])
             ->setData('store_id', $storeIds);
 
-        $this->pageRepository->save($page);
+        $this->entityManagement->getRepository(EntityManagement::TYPE_PAGE)->save($page);
 
         return true;
     }
@@ -329,7 +291,7 @@ class Import implements ContentImportInterface
      * Import a single block from an array and return false on error and true on success
      * @param array $blockData
      * @return bool
-     * @throws \Exception
+     * @throws Exception
      */
     public function importBlockContentFromArray(array $blockData): bool
     {
@@ -337,9 +299,14 @@ class Import implements ContentImportInterface
         $blockData = $this->loadBlocksByIdent($blockData, CmsBlockInterface::IDENTIFIER);
         $storeIds = $this->storeManagement->getStoreIdsByCodes($blockData['stores']);
 
-        $collection = $this->blockCollectionFactory->create();
-        $collection
-            ->addFieldToFilter(CmsBlockInterface::IDENTIFIER, $blockData['cms'][CmsBlockInterface::IDENTIFIER]);
+        $collection = $this->entityManagement
+            ->getRepository(EntityManagement::TYPE_BLOCK)
+            ->getList(
+                $this->searchCriteriaBuilder->addFilter(
+                    CmsBlockInterface::IDENTIFIER,
+                    $blockData['cms'][CmsBlockInterface::IDENTIFIER]
+                )->create()
+            )->getItems();
 
         $blockId = 0;
         foreach ($collection as $item) {
@@ -353,7 +320,7 @@ class Import implements ContentImportInterface
             }
         }
 
-        $block = $this->blockFactory->create();
+        $block = $this->entityManagement->getFactory(EntityManagement::TYPE_BLOCK);
         if ($blockId) {
             $block->load($blockId);
 
@@ -371,7 +338,7 @@ class Import implements ContentImportInterface
             ->setIsActive($cms[CmsBlockInterface::IS_ACTIVE])
             ->setData('store_id', $storeIds);
 
-        $this->blockRepository->save($block);
+        $this->entityManagement->getRepository(EntityManagement::TYPE_BLOCK)->save($block);
 
         return true;
     }
@@ -387,14 +354,17 @@ class Import implements ContentImportInterface
     {
         if (isset($cmsData['block_references'])) {
             $pairs = [];
+
+            $blockRepository = $this->entityManagement->getRepository(EntityManagement::TYPE_BLOCK);
+
             foreach ($cmsData['block_references'] as $blockId => $blockIdent) {
-                if (is_array($blockIdent)){
+                if (is_array($blockIdent)) {
                     foreach ($blockIdent as $blockIdentItem) {
-                        $block           = $this->blockRepository->getById($blockIdentItem);
+                        $block = $blockRepository->getById($blockIdentItem);
                         $pairs[$blockId] = $block->getId();
                     }
                 } else {
-                    $block           = $this->blockRepository->getById($blockIdent);
+                    $block = $blockRepository->getById($blockIdent);
                     $pairs[$blockId] = $block->getId();
                 }
             }
@@ -416,24 +386,46 @@ class Import implements ContentImportInterface
     }
 
     /**
-     * Set CMS mode
-     * @param $mode
-     * @return ContentImportInterface
+     * Import media data
+     *
+     * @param array $payload
+     * @param string $archivePath
+     * @param int $count
+     *
+     * @return int
+     * @throws FileSystemException
+     * @throws Exception
      */
-    public function setCmsModeOption($mode): ContentImportInterface
+    private function importMedia(array $payload, string $archivePath, int $count): int
     {
-        $this->cmsMode = $mode;
-        return $this;
-    }
+        // Import media
+        if ($archivePath && ($count > 0) && ($this->mediaMode != ContentImportInterface::OD_MEDIA_MODE_NONE)) {
+            if (isset($payload['media'])) {
+                foreach ($payload['media'] as $mediaFile) {
+                    $sourceFile = $archivePath . '/' . self::MEDIA_ARCHIVE_PATH . '/' . $mediaFile;
+                    $destFile = $this->fileManager->getMediaPath($mediaFile);
 
-    /**
-     * Set media mode
-     * @param $mode
-     * @return ContentImportInterface
-     */
-    public function setMediaModeOption($mode): ContentImportInterface
-    {
-        $this->mediaMode = $mode;
-        return $this;
+                    if ($this->file->fileExists($sourceFile, true)) {
+                        if ($this->file->fileExists($destFile, true) &&
+                            ($this->mediaMode == ContentImportInterface::OD_MEDIA_MODE_SKIP)
+                        ) {
+                            continue;
+                        }
+
+                        if (!$this->file->fileExists(dirname($destFile), false)) {
+                            if (!$this->file->mkdir(dirname($destFile))) {
+                                throw new Exception('Unable to create folder: ' . dirname($destFile));
+                            }
+                        }
+
+                        if (!$this->file->cp($sourceFile, $destFile)) {
+                            throw new Exception('Unable to save image: ' . $mediaFile);
+                        }
+                        $count++;
+                    }
+                }
+            }
+        }
+        return $count;
     }
 }