+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 09-07-2021
+## [1.1.0] - 10-06-2022
 ### Added
-- Possibility to import/export CMS pages or blocks.   
+- Possibility to import/export CMS pages or blocks.
+- Automatic deleting an old history backups
+- Refactored code